requirements.txt
amqp==2.3.2
apturl==0.5.2
argcomplete==1.8.1
asn1crypto==0.24.0
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
beautifulsoup4==4.6.0
billiard==3.5.0.4
bleach==2.1.4
blinker==1.4
boto3==1.4.4
botocore==1.5.95
Brlapi==0.6.6
celery==4.0.2
certifi==2017.7.27.1
chardet==3.0.4
chrome-gnome-shell==0.0.0
command-not-found==0.3
cryptography==2.1.4
cubes==1.1
cupshelpers==1.0
decorator==4.3.0
defer==1.0.6
defusedxml==0.5.0
distro-info==0.18
Django==2.1.1
django-activity-stream==0.6.3
django-bootstrap-ui==0.5.1
django-debug-toolbar==1.6
django-elasticsearch-dsl==0.5.0
django-extensions==1.7.8
django-filter==0.15.3
django-grappelli==2.8.1
django-redis==4.8.0
django-storages==1.5.2
django-tag-parser==2.1
djangorestframework==3.8.2
djangorestframework-jwt==1.8.0
docutils==0.14
dominate==2.1.17
elasticsearch==5.5.3
elasticsearch-dsl==5.4.0
entrypoints==0.2.3
expressions==0.2.3
feedparser==5.2.1
flake8==3.5.0
geopy==1.11.0
grako==3.99.9
greenlet==0.4.12
html5lib==0.9999999
httplib2==0.11.3
idna==2.6
inflection==0.3.1
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
isort==4.3.4
jedi==0.12.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
keyring==10.6.0
keyrings.alt==3.0
kombu==4.2.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.9.1
mccabe==0.6.1
mistune==0.8.3
more-itertools==4.3.0
msgpack==0.5.6
mysqlclient==1.3.10
nbconvert==5.4.0
nbformat==4.4.0
neovim==0.2.0
notebook==5.6.0
oauth==1.0.1
oauthlib==2.0.6
olefile==0.45.1
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
pluggy==0.7.1
progressbar==2.3
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.0.0
psycopg2==2.7.3.1
ptyprocess==0.6.0
py==1.6.0
pycairo==1.16.2
pycodestyle==2.3.1
pycrypto==2.6.1
pycups==1.9.73
pyflakes==1.6.0
Pygments==2.2.0
pygobject==3.26.1
PyJWT==1.6.4
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
pymacaroons==0.13.0
PyNaCl==1.1.2
PyPDF2==1.26.0
pyRFC3339==1.0
pyshp==1.2.3
pytest==3.8.0
pytest-django==3.1.2
python-apt==1.6.3
python-dateutil==2.7.3
python-debian==0.1.32
python-gnupg==0.4.1
pytz==2018.5
pyxdg==0.25
PyYAML==3.12
pyzmq==17.1.2
redis==2.10.5
reportlab==3.5.6
requests==2.10.0
requests-unixsocket==0.1.5
s3transfer==0.1.13
SecretStorage==2.3.1
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
SQLAlchemy==1.0.15
sqlparse==0.2.4
ssh-import-id==5.7
system-service==0.3
systemd-python==234
terminado==0.8.1
testpath==0.3.1
tornado==5.1
traitlets==4.3.2
typed-ast==1.1.0
ubuntu-drivers-common==0.0.0
Ubuntu-Make==18.5
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
unity-tweak-tool==0.0.7
urllib3==1.23
usb-creator==0.3.3
uWSGI==2.0.14
vine==1.1.4
virtualenv==16.0.0
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5
Werkzeug==0.12.2
widgetsnbextension==3.4.1
wrapt==1.10.11
xhtml2pdf==0.1b1
xkit==0.0.0
xlrd==0.9.4
zope.interface==4.3.2


